Staged changes

+++ backend/routes/posts.js
@@ -36,7 +36,8 @@ router.post("",
   const post = Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -67,17 +68,22 @@ router.put(
     imagePath: imagePath
   });
   console.log(post);
-  Post.updateOne({ _id: req.params.id }, post)
-    .then(post => {
-      res.status(201).json({
-        message: "Post modified successfully",
-        postId: post.id
-      });
-    }).catch(err => {
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
+    .then((post) => {
+      if(post.nModified > 0){
+        res.status(201).json({
+          message: "Post modified successfully",
+          postId: post.id,
+        });
+      } else {
+        res.status(401).json({ message: "Not authorized" });
+      }
+    })
+    .catch((err) => {
       console.log(err);
       res.status(404).json({
-        message: 'No post found like this.',
-        postId: req.params.id
+        message: "No post found like this.",
+        postId: req.params.id,
       });
     });
 });
@@ -121,8 +127,13 @@ router.delete(
   '/:id',
   checkAuth,
   (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    res.status(200).json({ message: "Post deleted" });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    console.log(result);
+    if(result.deletedCount > 0){
+      res.status(200).json({ message: "Post deleted" });
+    } else {
+      res.status(401).json({ message: "Not authorized" });
+    }
   });
 });
